+++ circuit/src/codec.rs
