+++ circuit/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod amounts;
 pub mod circuit;
+pub mod codec;
 pub mod exit_account;
-pub mod fcodec;
 pub mod gadgets;
 pub mod inputs;
 pub mod nullifier;
